Interpolate curved animation
AnimationContent behavior closer to XNA. This makes possible
to work with custom importers ported from/to XNA without
changing the importer or exporting again the FBX/animation.

XNA generates 60 keyframes per frame for every bone for
the full duration of the animation. For those keyframes it
seems that it samples/lerp the animation.
Keyframes in the FBX that don't fall on the right intervals
are dropped.